File Structure Refactor
Changed /private to /core as I added types to /core and /types is meant for internal and external use, so “private” is an inaccurate name. Added /types to core as several of the utils file use it. Also added basic.nim as the main arithmetic file, as having ../ imports bothers me.

New Structure:
* PMath.nim: Imports everything you need to get started
* PMath (All files are parts that use core files, and work together. Only basic is imported by default)
    * Core (The core utils that support the rest of the math)
        * utils, types
    * basic, extended etc.